frontend_api/requirements/base.txt
Django==5.1
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Pillow==10.0.0
requests==2.31.0
argon2-cffi==21.3.0
pytz==2023.3.post1
pika
faker
django-elasticsearch-dsl==7.3.0
django-elasticsearch-dsl-drf==0.22.5